Implement timers properly
The timers shown with `M-x timers` are now actual timings.

+++ src/dged/window.h
@@ -25,7 +25,7 @@ void windows_init(uint32_t height, uint32_t width,
 
 void windows_destroy();
 void windows_resize(uint32_t height, uint32_t width);
-void windows_update(void *(*frame_alloc)(size_t), uint64_t frame_time);
+void windows_update(void *(*frame_alloc)(size_t), float frame_time);
 void windows_render(struct display *display);
 
 struct window *root_window();
